Documentation, usability updates, and removing debug code.
+ First pass at documenting the Buffer class.
+ Starting Arena documentation.
+ Various usability updates, such as overloading operators.
+ Remove debug traces.

+++ Arena.cc
@@ -268,7 +268,8 @@ operator<<(std::ostream &os, Arena &arena)
 {
 	auto cursor = arena.Store();
 	char cursorString[33] = {0};
-	snprintf(cursorString, 32, "%#016llx", cursor);
+	snprintf(cursorString, 32, "%#016llx",
+			(long long unsigned int)cursor);
 
 	os << "Arena<";
 	switch (arena.Type()) {
